Added day 9 solutions
Rookie mistakes today. Part 1 took ages when I thought it was a good
idea to use a string to represent the tape. Then, I did part 2 not
focusing on speed, but it was enough to get top 1000 at least.

Don't look at that pointer mess for part 2. Although there are 5
pointers at some point, the code runs in O(n^2). It is possible to do it
linearly, as the block of equal characters have constant size (at most 10).

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  9   00:18:04   1657      0   00:33:54    838      0